agents/microsoft-agent-framework/requirements.txt
agent-framework-declarative>=0.3.0
python-dotenv>=1.0.1
